Simplify: extract transcode-settings group from api.js (slice 3)
Move getTranscodeSettings / setTranscodeSettings into src/api/transcode.js as
createTranscodeApi({ ctx }). The three videotoolbox-settings imports were used
only by these two methods, so they move into the new module and are dropped from
api.js (avoids an unused import). Closure dep is ctx only (metaDb persistence);
no this.* usage. api.js 4651 -> 4606 lines.
